==> 04.Sub_contig_anl.Rmd <==
---
title: "Sub-contig differentiation (i.e. fst analyses)"
author: "Jakub Kreisinger"
date: "9/15/2022"
output: html_document
---



# Packages, functions

```{r }
list.files("/media/kreising/DATA/data/BF_mysi/Fst_infekce_pro_Kub/")
library(lme4)
library(car)
library(ggeffects)
library(robustlmm)
library(heavy)
library(lqmm)
library(reshape2)
library(ggplot2)
library(nlmm)
library(ggplot2)
library(reshape2)
```

# Adjustments of input data for infected BULS

Here is example, how are the fst values coded by Popoolation2 sofware: 

                                  V1     V2 V3 V4 V5             V6             V7             V8             V9            V10            V11
1 NODE_1_length_863259_cov_11.829028 114458  1  1  4 1:2=0.25000000 1:3=0.00000000 1:4=0.00000000 2:3=0.25000000 2:4=0.29687500 3:4=0.00000000
2 NODE_1_length_863259_cov_11.829028 114461  1  1  4 1:2=0.25000000 1:3=0.00000000 1:4=0.00000000 2:3=0.25000000 2:4=0.29687500 3:4=0.00000000


Columns 1 - 2 correspond to the SNP IDs. From column 6 onwards, the FST values are given. Different pairs of "populations" are compared in each column. Here, wild donors and infected BULS from D-1, D2 and D7 were considered as separate populations.  


```{r }


dat<-read.delim("/media/kreising/DATA/data/BF_mysi/Fst_infekce_pro_Kub/BULS_infekce_nocontrol.count2cov4.fst",
                header = F)

# 
# head(dat)
names(dat)[6:11]<-c("INF_d0","INF_d2","INF_d7","d0_d2","d0_d7","d2_d7")

dat$INF_d0<-as.numeric(sapply(strsplit(dat$INF_d0, "="), function(x) x[2], simplify=T))
dat$INF_d2<-as.numeric(sapply(strsplit(dat$INF_d2, "="), function(x) x[2], simplify=T))
dat$INF_d7<-as.numeric(sapply(strsplit(dat$INF_d7, "="), function(x) x[2], simplify=T))
dat$d0_d2<-as.numeric(sapply(strsplit(dat$d0_d2, "="), function(x) x[2], simplify=T))
dat$d0_d7<-as.numeric(sapply(strsplit(dat$d0_d7, "="), function(x) x[2], simplify=T))
dat$d2_d7<-as.numeric(sapply(strsplit(dat$d2_d7, "="), function(x) x[2], simplify=T))

dat.sub<-dat[,6:11]
dat.sub$id<-paste(dat$V1,dat$V2,sep="~")

dat.sub.l<-melt(dat.sub, id.vars = c("id"))

dat.sub.l$node<-sapply(strsplit(dat.sub.l$id, "~"), function(x) x[1], simplify=T)
dat.sub.l$snp<-sapply(strsplit(dat.sub.l$id, "~"), function(x) x[2], simplify=T)

dat.sub.l_BULS<-dat.sub.l

```

# Adjustments of input data for infected Busna

The same as above for wild donors and infected BUSNA from each time point.

```{r }

dat<-read.delim("/media/kreising/DATA/data/BF_mysi/Fst_infekce_pro_Kub/BUSNA_infekce_nocontrol.count2cov4.fst",
                header = F)

names(dat)[6:11]<-c("INF_d0","INF_d2","INF_d7","d0_d2","d0_d7","d2_d7")

dat$INF_d0<-as.numeric(sapply(strsplit(dat$INF_d0, "="), function(x) x[2], simplify=T))
dat$INF_d2<-as.numeric(sapply(strsplit(dat$INF_d2, "="), function(x) x[2], simplify=T))
dat$INF_d7<-as.numeric(sapply(strsplit(dat$INF_d7, "="), function(x) x[2], simplify=T))
dat$d0_d2<-as.numeric(sapply(strsplit(dat$d0_d2, "="), function(x) x[2], simplify=T))
dat$d0_d7<-as.numeric(sapply(strsplit(dat$d0_d7, "="), function(x) x[2], simplify=T))
dat$d2_d7<-as.numeric(sapply(strsplit(dat$d2_d7, "="), function(x) x[2], simplify=T))

dat.sub<-dat[,6:11]
dat.sub$id<-paste(dat$V1,dat$V2,sep="~")

dat.sub.l<-melt(dat.sub, id.vars = c("id"))

dat.sub.l$node<-sapply(strsplit(dat.sub.l$id, "~"), function(x) x[1], simplify=T)
dat.sub.l$snp<-sapply(strsplit(dat.sub.l$id, "~"), function(x) x[2], simplify=T)
dat.sub.l_BUSNA<-dat.sub.l

```

# Adjustments of input data for contol BULS

The same as above for wild donors and control (non-infected) BULS from each time point.

```{r }

##########################
#BULS-control#############
##########################

dat<-read.delim("/media/kreising/DATA/data/BF_mysi/Fst_infekce_pro_Kub/BULS_infekce_onlycontrols.count2cov4.fst",
                header = F)

names(dat)[6:11]<-c("INF_d0","INF_d2","INF_d7","d0_d2","d0_d7","d2_d7")

dat$INF_d0<-as.numeric(sapply(strsplit(dat$INF_d0, "="), function(x) x[2], simplify=T))
dat$INF_d2<-as.numeric(sapply(strsplit(dat$INF_d2, "="), function(x) x[2], simplify=T))
dat$INF_d7<-as.numeric(sapply(strsplit(dat$INF_d7, "="), function(x) x[2], simplify=T))
dat$d0_d2<-as.numeric(sapply(strsplit(dat$d0_d2, "="), function(x) x[2], simplify=T))
dat$d0_d7<-as.numeric(sapply(strsplit(dat$d0_d7, "="), function(x) x[2], simplify=T))
dat$d2_d7<-as.numeric(sapply(strsplit(dat$d2_d7, "="), function(x) x[2], simplify=T))

dat.sub<-dat[,6:11]
dat.sub$id<-paste(dat$V1,dat$V2,sep="~")
dat.sub.l<-melt(dat.sub, id.vars = c("id"))
tapply(dat.sub.l$value,dat.sub.l$variable,mean)

dat.sub.l$node<-sapply(strsplit(dat.sub.l$id, "~"), function(x) x[1], simplify=T)
dat.sub.l$snp<-sapply(strsplit(dat.sub.l$id, "~"), function(x) x[2], simplify=T)
dat.sub.l_BULSc<-dat.sub.l

```

# Adjustments of input data for contol BULS

The same as above for wild donors and control (non-infected) BULS from each time point.
```{r }

##########################
#BUSNA-control#############
##########################

dat<-read.delim("/media/kreising/DATA/data/BF_mysi/Fst_infekce_pro_Kub/BUSNA_infekce_onlycontrols.count2cov4.fst",
                header = F)

names(dat)[6:11]<-c("INF_d0","INF_d2","INF_d7","d0_d2","d0_d7","d2_d7")

dat$INF_d0<-as.numeric(sapply(strsplit(dat$INF_d0, "="), function(x) x[2], simplify=T))
dat$INF_d2<-as.numeric(sapply(strsplit(dat$INF_d2, "="), function(x) x[2], simplify=T))
dat$INF_d7<-as.numeric(sapply(strsplit(dat$INF_d7, "="), function(x) x[2], simplify=T))
dat$d0_d2<-as.numeric(sapply(strsplit(dat$d0_d2, "="), function(x) x[2], simplify=T))
dat$d0_d7<-as.numeric(sapply(strsplit(dat$d0_d7, "="), function(x) x[2], simplify=T))
dat$d2_d7<-as.numeric(sapply(strsplit(dat$d2_d7, "="), function(x) x[2], simplify=T))

dat.sub<-dat[,6:11]
dat.sub$id<-paste(dat$V1,dat$V2,sep="~")
library(reshape2)
dat.sub.l<-melt(dat.sub, id.vars = c("id"))
library(ggplot2)
tapply(dat.sub.l$value,dat.sub.l$variable,mean)

dat.sub.l$node<-sapply(strsplit(dat.sub.l$id, "~"), function(x) x[1], simplify=T)
dat.sub.l$snp<-sapply(strsplit(dat.sub.l$id, "~"), function(x) x[2], simplify=T)

dat.sub.l_BUSNAc<-dat.sub.l

```

# Data.frame for statistical analyses

Only SNPs detected in both analyses comparing controls with infected recipients and donors with controls were considered. Subsequently, Fst values were categorised as "high" or "low" based on the median values. These steps were performed separately for the BULS and BUSNA strain.  

```{r }

exclude<-c("d0_d2","d0_d7","d2_d7" )
dat.sub.l_BUSNA<-dat.sub.l_BUSNA[!dat.sub.l_BUSNA$variable%in%exclude,]
dat.sub.l_BUSNAc<-dat.sub.l_BUSNAc[!dat.sub.l_BUSNAc$variable%in%exclude,]
dat.sub.l_BULS<-dat.sub.l_BULS[!dat.sub.l_BULS$variable%in%exclude,]
dat.sub.l_BULSc<-dat.sub.l_BULSc[!dat.sub.l_BULSc$variable%in%exclude,]

dat.sub.l_BUSNAc$type<-"control"
dat.sub.l_BUSNA$type<-"inf"
dat.sub.l_BULSc$type<-"control"
dat.sub.l_BULS$type<-"inf"

#intersect BUSNA
INTERSECT<-intersect(dat.sub.l_BUSNA$id,dat.sub.l_BUSNAc$id)
dat.sub.l_BUSNA.int<-dat.sub.l_BUSNA[dat.sub.l_BUSNA$id%in%INTERSECT,]
dat.sub.l_BUSNAc.int<-dat.sub.l_BUSNAc[dat.sub.l_BUSNAc$id%in%INTERSECT,]
dat.sub.l_BUSNAc.int$value.cat<-ifelse(dat.sub.l_BUSNAc.int$value>median(dat.sub.l_BUSNAc.int$value),1,0)
dat.sub.l_BUSNA.int$value.cat<-ifelse(dat.sub.l_BUSNA.int$value>median(dat.sub.l_BUSNA.int$value),1,0)

dat.sub.l_BUSNA.all<-rbind(dat.sub.l_BUSNA.int,dat.sub.l_BUSNAc.int)

#intersect BULS
INTERSECT<-intersect(dat.sub.l_BULS$id,dat.sub.l_BULSc$id)
dat.sub.l_BULS.int<-dat.sub.l_BULS[dat.sub.l_BULS$id%in%INTERSECT,]
dat.sub.l_BULSc.int<-dat.sub.l_BULSc[dat.sub.l_BULSc$id%in%INTERSECT,]
dat.sub.l_BULS.int$value.cat<-ifelse(dat.sub.l_BULS.int$value>median(dat.sub.l_BULS.int$value),1,0)
dat.sub.l_BULSc.int$value.cat<-ifelse(dat.sub.l_BULSc.int$value>median(dat.sub.l_BULSc.int$value),1,0)

dat.sub.l_BULS.all<-rbind(dat.sub.l_BULS.int,dat.sub.l_BULSc.int)

#merge data.frames
dat.sub.l_BUSNA.all$Strain<-"BUSNA"
dat.sub.l_BULS.all$Strain<-"BULS"
ALL_data<-rbind(dat.sub.l_BUSNA.all,dat.sub.l_BULS.all)

```

# Statistical models

The effect of infection on the variation of FST values (categorized as high or low) was analyzed using mixed models with binomial distribution. SNP id was considered as a random effect. Due to the different effects of infection treatment between mouse mouse stains, separate models were fitted for BULS and BUSNA. In the first step, the effect of infection treatment was tested with a model that included the interaction between the treatment group ("variable" in the model equation) and the experimental time point ("type" in the model equation). Subsequently, separate models were fitted for uninfected controls and infected recipients.  

## BULS subset
```{r }

ALL_data.BULS<-ALL_data[ALL_data$Strain=="BULS",]
tapply(dat.sub.l_BULS.all$value.cat,list(dat.sub.l_BULS.all$variable,dat.sub.l_BULS.all$type),mean)

model.BULS.1<-glmer(value.cat~type+variable+
                   type:variable+
                   (1|id),binomial,data = dat.sub.l_BULS.all)

model.BULS.2<-glmer(value.cat~type+variable+
                   # type:variable+
                   (1|id),binomial,data = dat.sub.l_BULS.all)

summary(model.BULS.1)
anova(model.BULS.1,model.BULS.2)

###Infected
ALL_data.BULS.inf<-ALL_data.BULS[dat.sub.l_BULS.all$type!="control",]
model.BULS.infected.1<-glmer(value.cat~variable+
                             (1|id),binomial,data = ALL_data.BULS.inf)
model.BULS.infected.2<-glmer(value.cat~1+
                             (1|id),binomial,data = ALL_data.BULS.inf)
model.BULS.infected.null<-glmer(value.cat~variable-1+
                             (1|id),binomial,data = ALL_data.BULS.inf)

summary(model.BULS.infected.1)
anova(model.BULS.infected.1,model.BULS.infected.2)

summary(model.BULS.infected.1)

###Controls
ALL_data.BULS.cont<-ALL_data.BULS[dat.sub.l_BULS.all$type=="control",]
model.BULS.control.1<-glmer(value.cat~variable+
                             (1|id),binomial,data = ALL_data.BULS.cont)
model.BULS.control.2<-glmer(value.cat~1+
                             (1|id),binomial,data = ALL_data.BULS.cont)
model.BULS.control.null<-glmer(value.cat~variable-1+
                             (1|id),binomial,data = ALL_data.BULS.cont)

summary(model.BULS.control.1)
anova(model.BULS.control.1,model.BULS.control.2)

summary(model.BULS.control.1)

```

## BUSNA subset
```{r }

ALL_data.BUSNA<-ALL_data[ALL_data$Strain=="BUSNA",]
tapply(ALL_data.BUSNA$value.cat,list(ALL_data.BUSNA$variable,ALL_data.BUSNA$type),mean)

model.BUSNA.1<-glmer(value.cat~type+variable+
                   type:variable+
                   (1|id),binomial,data = ALL_data.BUSNA)

model.BUSNA.2<-glmer(value.cat~type+variable+
                   # type:variable+
                   (1|id),binomial,data = ALL_data.BUSNA)

summary(model.BUSNA.1)
anova(model.BUSNA.1,model.BUSNA.2)

###Infected
ALL_data.BUSNA.inf<-ALL_data.BUSNA[ALL_data.BUSNA$type!="control",]
model.BUSNA.infected.1<-glmer(value.cat~variable+
                             (1|id),binomial,data = ALL_data.BUSNA.inf)
model.BUSNA.infected.2<-glmer(value.cat~1+
                             (1|id),binomial,data = ALL_data.BUSNA.inf)
model.BUSNA.infected.null<-glmer(value.cat~variable-1+
                             (1|id),binomial,data = ALL_data.BUSNA.inf)

summary(model.BUSNA.infected.1)
anova(model.BUSNA.infected.1,model.BUSNA.infected.2)

summary(model.BUSNA.infected.1)

###Controls
ALL_data.BUSNA.cont<-ALL_data.BUSNA[ALL_data.BUSNA$type=="control",]
model.BUSNA.control.1<-glmer(value.cat~variable+
                             (1|id),binomial,data = ALL_data.BUSNA.cont)
model.BUSNA.control.2<-glmer(value.cat~1+
                             (1|id),binomial,data = ALL_data.BUSNA.cont)
model.BUSNA.control.null<-glmer(value.cat~variable-1+
                             (1|id),binomial,data = ALL_data.BUSNA.cont)

summary(model.BUSNA.control.1)
anova(model.BUSNA.control.1,model.BUSNA.control.2)

summary(model.BUSNA.control.1)


```


# Model predictions and graphical outputs

```{r }
#intersect BULS
INTERSECT<-intersect(dat.sub.l_BULS$id,dat.sub.l_BULSc$id)
dat.sub.l_BULS.int<-dat.sub.l_BULS[dat.sub.l_BULS$id%in%INTERSECT,]
dat.sub.l_BULSc.int<-dat.sub.l_BULSc[dat.sub.l_BULSc$id%in%INTERSECT,]

dat.sub.l_BULS.int$type<-gsub("inf","infected",dat.sub.l_BULS.int$type)
dat.sub.l_BULSc.int$type<-gsub("inf","infected",dat.sub.l_BULSc.int$type)

#q50
dat.sub.l_BULS.int$value.cat<-ifelse(dat.sub.l_BULS.int$value>quantile(dat.sub.l_BULS.int$value,0.5),1,0)
dat.sub.l_BULSc.int$value.cat<-ifelse(dat.sub.l_BULSc.int$value>quantile(dat.sub.l_BULSc.int$value,0.5),1,0)

dat.sub.l_BULS.all<-rbind(dat.sub.l_BULS.int,dat.sub.l_BULSc.int)

dat.sub.l_BULS.all$variable<-as.character(dat.sub.l_BULS.all$variable)
dat.sub.l_BULS.all$variable<-gsub("INF_d","D",dat.sub.l_BULS.all$variable)
dat.sub.l_BULS.all$variable<-gsub("D0","D-1",dat.sub.l_BULS.all$variable)


model.BULS.1<-glmer(value.cat~type+variable+
                      type:variable+
                      (1|id),binomial,data = dat.sub.l_BULS.all)

pred<-ggpredict(model.BULS.1,terms = c("variable", "type"))
q50.buls<-plot(pred)

#intersect BULS
INTERSECT<-intersect(dat.sub.l_BUSNA$id,dat.sub.l_BUSNAc$id)
dat.sub.l_BUSNA.int<-dat.sub.l_BUSNA[dat.sub.l_BUSNA$id%in%INTERSECT,]
dat.sub.l_BUSNAc.int<-dat.sub.l_BUSNAc[dat.sub.l_BUSNAc$id%in%INTERSECT,]

dat.sub.l_BUSNA.int$type<-gsub("inf","infected",dat.sub.l_BUSNA.int$type)
dat.sub.l_BUSNAc.int$type<-gsub("inf","infected",dat.sub.l_BUSNAc.int$type)

#q50
dat.sub.l_BUSNA.int$value.cat<-ifelse(dat.sub.l_BUSNA.int$value>quantile(dat.sub.l_BUSNA.int$value,0.5),1,0)
dat.sub.l_BUSNAc.int$value.cat<-ifelse(dat.sub.l_BUSNAc.int$value>quantile(dat.sub.l_BUSNAc.int$value,0.5),1,0)

dat.sub.l_BUSNA.all<-rbind(dat.sub.l_BUSNA.int,dat.sub.l_BUSNAc.int)

dat.sub.l_BUSNA.all$variable<-as.character(dat.sub.l_BUSNA.all$variable)
dat.sub.l_BUSNA.all$variable<-gsub("INF_d","D",dat.sub.l_BUSNA.all$variable)
dat.sub.l_BUSNA.all$variable<-gsub("D0","D-1",dat.sub.l_BUSNA.all$variable)

model.BUSNA.1<-glmer(value.cat~type+variable+
                      type:variable+
                      (1|id),binomial,data = dat.sub.l_BUSNA.all)

pred<-ggpredict(model.BUSNA.1,terms = c("variable", "type"))
q50.busna<-plot(pred)


q50.busna<-q50.busna+theme_bw()+scale_color_discrete(labels=c('Donors-Controls', 'Donors-Recipients'))
q50.buls<-q50.buls+theme_bw()+scale_color_discrete(labels=c('Donors-Controls', 'Donors-Recipients'))
# q50.busna+scale_color_manual(name = "type",values = c("Donors-Controls"="red","Donors-Recipients"="blue"))

library(ggpubr)
library(scales)
q50.busna.m<-
q50.busna+theme(axis.title.x = element_blank(),
                axis.title.y = element_text(size = 8),
                plot.title = element_text(size=8),
                axis.text.x = element_text(size = 8),
                axis.text.y = element_text(size = 8),
                legend.title = element_blank())

q50.busna.m<-
q50.busna.m+ggtitle("B) BUSNA")+ylab("High-Fst SNPs\n(proportion)")+scale_y_continuous(labels = waiver())


q50.buls.m<-
q50.buls+theme(axis.title.x = element_blank(),
                axis.title.y = element_text(size = 8),
                plot.title = element_text(size=8),
                axis.text.x = element_text(size = 8),
                axis.text.y = element_text(size = 8),
                legend.title = element_blank())
  

q50.buls.m<-
q50.buls.m+ggtitle("A) BULS")+ylab("High-Fst SNPs\n(proportion)")+
  scale_y_continuous(labels = waiver())

GRID<-ggarrange(q50.buls.m,
                q50.busna.m,
                nrow = 2,
                common.legend = T,
                legend = "bottom")

ggsave(GRID,filename = "/media/kreising/DATA/data/BF_mysi/RESULTS_10.6.2024/FST_ggplot_binomial_NEW.pdf",
       width = 8,height = 12,units = "cm")

GRID
```
